Refactor debug console into own module
