+++ qt/aqt/widgetgallery.py
@@ -3,7 +3,7 @@
 
 import aqt
 import aqt.main
-from aqt.qt import QDialog, qconnect
+from aqt.qt import QDialog, QWidget, qconnect
 from aqt.theme import WidgetStyle
 from aqt.utils import restoreGeom, saveGeom
 
@@ -11,9 +11,9 @@ from aqt.utils import restoreGeom, saveGeom
 class WidgetGallery(QDialog):
     silentlyClose = True
 
-    def __init__(self, mw: aqt.main.AnkiQt) -> None:
-        super().__init__(mw)
-        self.mw = mw.weakref()
+    def __init__(self, parent: QWidget) -> None:
+        assert aqt.mw
+        super().__init__(parent)
 
         self.form = aqt.forms.widgets.Ui_Dialog()
         self.form.setupUi(self)
@@ -29,10 +29,10 @@ class WidgetGallery(QDialog):
         self.form.styleComboBox.addItems(
             [member.name.lower().capitalize() for member in WidgetStyle]
         )
-        self.form.styleComboBox.setCurrentIndex(self.mw.pm.get_widget_style())
+        self.form.styleComboBox.setCurrentIndex(aqt.mw.pm.get_widget_style())
         qconnect(
             self.form.styleComboBox.currentIndexChanged,
-            self.mw.pm.set_widget_style,
+            aqt.mw.pm.set_widget_style,
         )
 
     def reject(self) -> None:
